requirements.txt
fastapi
uvicorn
requests
pydantic
streamlit
keybert
nltk
graphviz
python-dotenv
openai